+++ packages/apollo-angular-link-http-batch/tests/HttpBatchLink.spec.ts
@@ -76,6 +76,32 @@ describe('HttpBatchLink', () => {
     }, 50);
   });
 
+  test('should fail on uri as function', (done: jest.DoneCallback) => {
+    const link = httpLink.create({
+      uri: () => '/graphql',
+    });
+    const op = {
+      query: gql`
+        query heroes {
+          heroes {
+            name
+          }
+        }
+      `,
+      operationName: 'heroes',
+      variables: {},
+    };
+
+    execute(link, op).subscribe({
+      next: () => {
+        done.fail('Should not be here');
+      },
+      error: () => {
+        done();
+      },
+    });
+  });
+
   test('should support multiple queries', (done: jest.DoneCallback) => {
     const link = httpLink.create({uri: 'graphql', batchKey: () => 'bachKey'});
     const op1 = {