+++ packages/apollo-angular-link-http/ng-package.js
@@ -4,8 +4,8 @@ module.exports = {
   dest: 'build',
   lib: {
     entryFile: 'src/index.ts',
-    flatModuleFile: 'ng.apolloLink.http',
+    flatModuleFile: 'ngApolloLinkHttp',
     umdModuleIds,
   },
-  whitelistedNonPeerDependencies: ['apollo-angular-link-http-common'],
+  whitelistedNonPeerDependencies: ['apollo-angular-link-http-common', 'extract-files'],
 };