+++ packages/apollo-angular/src/Apollo.ts
@@ -2,7 +2,6 @@ import {Injectable, Optional, Inject, NgZone} from '@angular/core';
 import {
   ApolloClient,
   QueryOptions,
-  WatchQueryOptions,
   MutationOptions,
   ApolloQueryResult,
   SubscriptionOptions,
@@ -13,20 +12,26 @@ import {FetchResult} from 'apollo-link';
 import {Observable, from} from 'rxjs';
 
 import {QueryRef} from './QueryRef';
-import {ExtraSubscriptionOptions, R} from './types';
-import {APOLLO_OPTIONS} from './tokens';
+import {
+  WatchQueryOptions,
+  ExtraSubscriptionOptions,
+  R,
+  NamedOptions,
+} from './types';
+import {APOLLO_OPTIONS, APOLLO_NAMED_OPTIONS} from './tokens';
 import {fromPromise, wrapWithZone, fixObservable} from './utils';
 
 export class ApolloBase<TCacheShape = any> {
   constructor(
-    private ngZone: NgZone,
-    private _client?: ApolloClient<TCacheShape>,
+    protected ngZone: NgZone,
+    protected _client?: ApolloClient<TCacheShape>,
   ) {}
 
   public watchQuery<T, V = R>(options: WatchQueryOptions<V>): QueryRef<T, V> {
     return new QueryRef<T, V>(
       this.client.watchQuery<T, V>({...options}) as ObservableQuery<T, V>,
       this.ngZone,
+      options,
     );
   }
 
@@ -49,7 +54,7 @@ export class ApolloBase<TCacheShape = any> {
   public subscribe<T, V = R>(
     options: SubscriptionOptions<V>,
     extra?: ExtraSubscriptionOptions,
-  ): Observable<any> {
+  ): Observable<FetchResult<T>> {
     const obs = from(fixObservable(this.client.subscribe<T, V>({...options})));
 
     return extra && extra.useZone !== true
@@ -57,10 +62,19 @@ export class ApolloBase<TCacheShape = any> {
       : wrapWithZone(obs, this.ngZone);
   }
 
+  /**
+   * Get an access to an instance of ApolloClient
+   */
   public getClient() {
     return this._client;
   }
 
+  /**
+   * Set a new instance of ApolloClient
+   * Remember to clean up the store before setting a new client.
+   *
+   * @param client ApolloClient instance
+   */
   public setClient(client: ApolloClient<TCacheShape>) {
     if (this._client) {
       throw new Error('Client has been already defined');
@@ -98,36 +112,64 @@ export class Apollo extends ApolloBase<any> {
     @Optional()
     @Inject(APOLLO_OPTIONS)
     apolloOptions?: ApolloClientOptions<any>,
+    @Optional()
+    @Inject(APOLLO_NAMED_OPTIONS)
+    apolloNamedOptions?: NamedOptions,
   ) {
     super(_ngZone);
 
     if (apolloOptions) {
       this.createDefault(apolloOptions);
     }
+
+    if (apolloNamedOptions && typeof apolloNamedOptions === 'object') {
+      for (const name in apolloNamedOptions) {
+        if (apolloNamedOptions.hasOwnProperty(name)) {
+          const options = apolloNamedOptions[name];
+          this.createNamed(name, options);
+        }
+      }
+    }
   }
 
+  /**
+   * Create an instance of ApolloClient
+   * @param options Options required to create ApolloClient
+   * @param name client's name
+   */
   public create<TCacheShape>(
     options: ApolloClientOptions<TCacheShape>,
     name?: string,
   ): void {
-    if (name && name !== 'default') {
-      this.createNamed<TCacheShape>(name, options);
-    } else {
+    if (isDefault(name)) {
       this.createDefault<TCacheShape>(options);
+    } else {
+      this.createNamed<TCacheShape>(name, options);
     }
   }
 
+  /**
+   * Use a default ApolloClient
+   */
   public default(): ApolloBase<any> {
     return this;
   }
 
+  /**
+   * Use a named ApolloClient
+   * @param name client's name
+   */
   public use(name: string): ApolloBase<any> {
-    if (name === 'default') {
+    if (isDefault(name)) {
       return this.default();
     }
     return this.map.get(name);
   }
 
+  /**
+   * Create a default ApolloClient, same as `apollo.create(options)`
+   * @param options ApolloClient's options
+   */
   public createDefault<TCacheShape>(
     options: ApolloClientOptions<TCacheShape>,
   ): void {
@@ -138,6 +180,11 @@ export class Apollo extends ApolloBase<any> {
     return this.setClient(new ApolloClient<TCacheShape>(options));
   }
 
+  /**
+   * Create a named ApolloClient, same as `apollo.create(options, name)`
+   * @param name client's name
+   * @param options ApolloClient's options
+   */
   public createNamed<TCacheShape>(
     name: string,
     options: ApolloClientOptions<TCacheShape>,
@@ -150,4 +197,20 @@ export class Apollo extends ApolloBase<any> {
       new ApolloBase(this._ngZone, new ApolloClient<TCacheShape>(options)),
     );
   }
+
+  /**
+   * Remember to clean up the store before removing a client
+   * @param name client's name
+   */
+  public removeClient(name?: string): void {
+    if (isDefault(name)) {
+      this._client = undefined;
+    } else {
+      this.map.delete(name);
+    }
+  }
+}
+
+function isDefault(name?: string): boolean {
+  return !name || name === 'default';
 }