+++ packages/apollo-angular/src/types.ts
@@ -3,6 +3,7 @@ import {
   QueryOptions as CoreQueryOptions,
   MutationOptions as CoreMutationOptions,
   SubscriptionOptions as CoreSubscriptionOptions,
+  ApolloClientOptions,
 } from 'apollo-client';
 import {ExecutionResult} from 'graphql';
 
@@ -16,18 +17,30 @@ export interface ExtraSubscriptionOptions {
 
 export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
-export interface WatchQueryOptions<V>
-  extends Omit<CoreWatchQueryOptions<V>, 'query' | 'variables'> {}
+export interface WatchQueryOptionsAlone<V>
+  extends Omit<WatchQueryOptions<V>, 'query' | 'variables'> {}
 
-export interface QueryOptions<V>
+export interface QueryOptionsAlone<V>
   extends Omit<CoreQueryOptions<V>, 'query' | 'variables'> {}
 
-export interface MutationOptions<T, V>
+export interface MutationOptionsAlone<T, V>
   extends Omit<CoreMutationOptions<T, V>, 'mutation' | 'variables'> {}
 
-export interface SubscriptionOptions<V>
+export interface SubscriptionOptionsAlone<V>
   extends Omit<CoreSubscriptionOptions<V>, 'query' | 'variables'> {}
+
+export interface WatchQueryOptions<V> extends CoreWatchQueryOptions<V> {
+  /**
+   * Observable starts with `{ loading: true }`.
+   * There's a big chance the next major version will enable that by default.
+   *
+   * Disabled by default
+   */
+  useInitialLoading?: boolean;
+}
 
 export interface SubscriptionResult<T> extends ExecutionResult {
   data?: T;
 }
+
+export type NamedOptions = Record<string, ApolloClientOptions<any>>;