+++ packages/apollo-angular/src/utils.ts
@@ -1,6 +1,4 @@
 import {NgZone} from '@angular/core';
-import {ObservableQuery} from 'apollo-client';
-import {Observable as ApolloObservable} from 'apollo-client/util/Observable';
 import {observeOn} from 'rxjs/operators';
 import {
   Observable,
@@ -50,9 +48,7 @@ export class ZoneScheduler implements SchedulerLike {
 // XXX: Apollo's QueryObservable is not compatible with RxJS
 // TODO: remove it in one of future releases
 // https://github.com/ReactiveX/rxjs/blob/9fb0ce9e09c865920cf37915cc675e3b3a75050b/src/internal/util/subscribeTo.ts#L32
-export function fixObservable<T>(
-  obs: ObservableQuery<T> | ApolloObservable<T>,
-): ObservableQuery<T> | ApolloObservable<T> {
+export function fixObservable<T>(obs: T): T {
   (obs as any)[observable] = () => obs;
   return obs;
 }