+++ packages/apollo-angular/testing/src/backend.ts
@@ -1,15 +1,11 @@
 import {Injectable} from '@angular/core';
 import {Observer} from 'rxjs';
-import {
-  Operation,
-  FetchResult,
-  Observable as LinkObservable,
-} from 'apollo-link';
+import {FetchResult, Observable as LinkObservable} from 'apollo-link';
 import {DocumentNode} from 'graphql';
-import {print} from 'graphql/language/printer';
+import {print} from 'graphql';
 
 import {ApolloTestingController, MatchOperation} from './controller';
-import {TestOperation} from './operation';
+import {TestOperation, Operation} from './operation';
 
 /**
  * A testing backend for `Apollo`.
@@ -113,9 +109,7 @@ export class ApolloTestingBackend implements ApolloTestingController {
     const matches = this.match(match);
     if (matches.length > 1) {
       throw new Error(
-        `Expected one matching operation for criteria "${description}", found ${
-          matches.length
-        } operations.`,
+        `Expected one matching operation for criteria "${description}", found ${matches.length} operations.`,
       );
     }
     if (matches.length === 0) {
@@ -135,9 +129,7 @@ export class ApolloTestingBackend implements ApolloTestingController {
     const matches = this.match(match);
     if (matches.length > 0) {
       throw new Error(
-        `Expected zero matching operations for criteria "${description}", found ${
-          matches.length
-        }.`,
+        `Expected zero matching operations for criteria "${description}", found ${matches.length}.`,
       );
     }
   }