+++ packages/apollo-angular/testing/src/module.ts
@@ -1,15 +1,38 @@
 import {ApolloModule, Apollo} from 'apollo-angular';
-import {ApolloLink} from 'apollo-link';
+import {ApolloLink, Operation as LinkOperation} from 'apollo-link';
 import {InMemoryCache} from 'apollo-cache-inmemory';
 import {ApolloCache} from 'apollo-cache';
-import {NgModule, InjectionToken, Inject, Optional} from '@angular/core';
+import {
+  NgModule,
+  InjectionToken,
+  Inject,
+  Optional,
+  ModuleWithProviders,
+} from '@angular/core';
 
 import {ApolloTestingController} from './controller';
 import {ApolloTestingBackend} from './backend';
+import {Operation} from './operation';
+
+export type NamedCaches = Record<string, ApolloCache<any> | undefined | null>;
 
 export const APOLLO_TESTING_CACHE = new InjectionToken<ApolloCache<any>>(
   'apollo-angular/testing cache',
 );
+
+export const APOLLO_TESTING_NAMED_CACHE = new InjectionToken<NamedCaches>(
+  'apollo-angular/testing named cache',
+);
+
+export const APOLLO_TESTING_CLIENTS = new InjectionToken<string[]>(
+  'apollo-angular/testing named clients',
+);
+
+function addClient(name: string, op: LinkOperation): Operation {
+  (op as Operation).clientName = name;
+
+  return op as Operation;
+}
 
 @NgModule({
   imports: [ApolloModule],
@@ -18,21 +41,59 @@ export const APOLLO_TESTING_CACHE = new InjectionToken<ApolloCache<any>>(
     {provide: ApolloTestingController, useExisting: ApolloTestingBackend},
   ],
 })
-export class ApolloTestingModule {
+export class ApolloTestingModuleCore {
   constructor(
     apollo: Apollo,
     backend: ApolloTestingBackend,
     @Optional()
+    @Inject(APOLLO_TESTING_CLIENTS)
+    namedClients?: string[],
+    @Optional()
     @Inject(APOLLO_TESTING_CACHE)
     cache?: ApolloCache<any>,
+    @Optional()
+    @Inject(APOLLO_TESTING_NAMED_CACHE)
+    namedCaches?: any, // FIX: using NamedCaches here makes ngc fail
   ) {
-    apollo.create({
-      link: new ApolloLink(operation => backend.handle(operation)),
-      cache:
-        cache ||
-        new InMemoryCache({
-          addTypename: false,
-        }),
-    });
+    function createOptions(name: string, c?: ApolloCache<any> | null) {
+      return {
+        link: new ApolloLink(operation =>
+          backend.handle(addClient(name, operation)),
+        ),
+        cache:
+          c ||
+          new InMemoryCache({
+            addTypename: false,
+          }),
+      };
+    }
+
+    apollo.create(createOptions('default', cache));
+
+    if (namedClients && namedClients.length) {
+      namedClients.forEach(name => {
+        const caches =
+          namedCaches && typeof namedCaches === 'object' ? namedCaches : {};
+
+        apollo.createNamed(name, createOptions(name, caches[name]));
+      });
+    }
+  }
+}
+
+@NgModule({
+  imports: [ApolloTestingModuleCore],
+})
+export class ApolloTestingModule {
+  static withClients(names: string[]): ModuleWithProviders {
+    return {
+      ngModule: ApolloTestingModuleCore,
+      providers: [
+        {
+          provide: APOLLO_TESTING_CLIENTS,
+          useValue: names,
+        },
+      ],
+    };
   }
 }