+++ packages/apollo-angular-link-http-batch/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change log
 
+### v1.7.0
+
+- Sync with common@1.8.0
+
+### v1.6.0
+
+- Support Angular 8
+
+### v1.5.0
+
+- Make sure people don't use `uri` as function [PR #1084](https://github.com/apollographql/apollo-angular/pull/1084)
+
 ### v1.4.0
 
 - Angular 7 [PR #913](https://github.com/apollographql/apollo-angular/pull/913)